use StateFlow instead of LiveData in viewModels

+++ app/src/main/java/com/vendetta/weather/data/network/dto/LocationDto.kt
@@ -26,7 +26,7 @@ data class LocationDto(
     val tzId: String,
     @SerializedName("localtime_epoch")
     @Expose
-    val localtimeEpoch: Int,
+    val localtimeEpoch: Long,
     @SerializedName("localtime")
     @Expose
     val localtime: String
